refactor: removed balance and contracts page components
