--- src/components/balance/modals/ModalCustomSignTx.vue
@@ -1,1 +1,0 @@
-// @TODO-bobo : packages/page-custom-signature/src/EcdsaCallSigner/CustomSignTx.tsx
