+++ src/components/dapp-staking/modals/InputFile.vue
