+++ src/components/dapp-staking/modals/StakeModal.vue
@@ -58,10 +58,10 @@
 
 <script lang="ts">
 import BN from 'bn.js';
-import FormatBalance from 'components/balance/FormatBalance.vue';
-import ModalSelectAccount from 'components/balance/modals/ModalSelectAccount.vue';
+import FormatBalance from 'components/common/FormatBalance.vue';
+import ModalSelectAccount from 'components/common/ModalSelectAccount.vue';
 import Modal from 'components/common/Modal.vue';
-import { Role } from 'src/components/balance/modals/ModalTransferAmount.vue';
+import { Role } from 'src/components/common/ModalTransferAmount.vue';
 import Avatar from 'src/components/common/Avatar.vue';
 import Button from 'src/components/common/Button.vue';
 import InputAmount from 'src/components/common/InputAmount.vue';
