+++ src/components/header/modals/ModalInstallWallet.vue
