+++ src/router/routes.ts
@@ -3,7 +3,6 @@ import { RouteRecordRaw } from 'vue-router';
 import Assets from 'pages/Assets.vue';
 import Store from 'src/pages/DappStaking.vue';
 import Dashboard from 'src/pages/Dashboard.vue';
-import ConnectWallet from 'src/components/balance/ConnectWallet.vue';
 import DiscoverDappsTab from 'components/dapp-staking/DiscoverDappsTab.vue';
 import ManageDappsTab from 'components/dapp-staking/ManageDappsTab.vue';
 import EvmWidget from 'components/bridge/evm/EvmWidget.vue';
@@ -17,32 +16,6 @@ const routes: RouteRecordRaw[] = [
   {
     path: '/store/discover-dapps',
     redirect: '/dapp-staking/discover',
-  },
-  {
-    path: '/balance',
-    name: 'Balance',
-    redirect: '/assets',
-    children: [
-      {
-        path: '',
-        redirect: '/assets',
-      },
-      {
-        path: 'balance-plasm',
-        redirect: '/assets',
-      },
-      {
-        path: 'wallet',
-        component: ConnectWallet,
-        redirect: '/assets',
-        children: [
-          {
-            path: 'deeplink-metamask',
-            redirect: '/assets',
-          },
-        ],
-      },
-    ],
   },
   {
     path: '/assets',
@@ -60,25 +33,6 @@ const routes: RouteRecordRaw[] = [
     name: 'Dashboard',
     component: Dashboard,
   },
-  // {
-  //   path: '/contracts',
-  //   name: 'Contracts',
-  //   component: DApps,
-  //   children: [
-  //     {
-  //       path: '',
-  //       redirect: '/contracts/create-contract',
-  //     },
-  //     {
-  //       path: 'create-dapps',
-  //       redirect: '/contracts/create-contract',
-  //     },
-  //     {
-  //       path: 'create-contract',
-  //       component: CreateDappsTab,
-  //     },
-  //   ],
-  // },
   {
     path: '/dapp-staking',
     name: 'dApp Staking',
